Modificação na listagem dos alarmes (sem edição)

+++ app/src/main/java/com/adht/android/medicontrol/alarme/negocio/AlarmeServices.java
@@ -24,7 +24,7 @@ public class AlarmeServices {
         return daoAlarme.listar(idPaciente);
     }
 
-    public void atualizar(Alarme alarme, int idAlarme) throws IOException {
+    public void atualizar(Alarme alarme, long idAlarme) throws IOException {
         AlarmeDAOSQLite dao = new AlarmeDAOSQLite();
         dao.atualizar(alarme, idAlarme);
     }
